Move V2 parquet format classes to shared format package
Relocate DeltaParquetFileFormatV2 and ProtocolMetadataAdapterV2 from the
read package into a shared format package and update all import sites.
This is a mechanical refactor with no behavior changes.
